requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Development
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0
